api/requirements.txt
azure-functions
sib-api-v3-sdk